ts書き出し以外のレビュー
